feat: Add function to generate unique random numbers (#9)
This commit adds a new function `generateUniqueRandomNumbers` to generate a specified number of unique random numbers within a given range. The function checks for the edge case where the count of unique numbers is greater than the range size and throws an error. It then generates the unique random numbers using a set and returns the result. This function is used in the `draw` function to ensure that participants are selected randomly and without repetition.